Fixed all the stuff
