+++ AppletRunner.html
@@ -2,7 +2,7 @@
 <!-- This is so I can run the program from MS-DOS, if anyone is using a Mac,
 could you test if you can run Graphics.class without this html file? -->
 <p> This file launches the Mouse.java Applet</p>
-<applet code="Mouse.class" height=800 width=800>
+<applet code="GraphicThing.class" height=800 width=800>
 </applet>
 </html>
